refact : 그룹 세부사항 조회 서비스 수정
- 쿼리 주최를 변경함
