+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group/repository/GroupQueryRepository.java
@@ -1,7 +1,6 @@
 package site.timecapsulearchive.core.domain.group.repository;
 
 import java.time.ZonedDateTime;
-import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Slice;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupDetailDto;
@@ -16,11 +15,8 @@ public interface GroupQueryRepository {
         final ZonedDateTime createdAt
     );
 
-    Optional<GroupDetailDto> findGroupDetailByGroupIdAndMemberId(final Long groupId, final Long memberId);
-
-    Long findGroupCapsuleCount(final Long groupId);
+    Optional<GroupDetailDto> findGroupDetailByGroupIdAndMemberId(final Long groupId,
+        final Long memberId);
 
     Boolean findGroupEditPermission(final Long groupId, final Long memberId);
-
-    List<Long> findFriendIds(final List<Long> groupMemberIds, final Long memberId);
 }
