+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group/repository/GroupQueryRepositoryImpl.java
@@ -2,8 +2,6 @@ package site.timecapsulearchive.core.domain.group.repository;
 
 import static com.querydsl.core.group.GroupBy.groupBy;
 import static com.querydsl.core.group.GroupBy.list;
-import static site.timecapsulearchive.core.domain.capsule.entity.QCapsule.capsule;
-import static site.timecapsulearchive.core.domain.friend.entity.QMemberFriend.memberFriend;
 import static site.timecapsulearchive.core.domain.group.entity.QGroup.group;
 import static site.timecapsulearchive.core.domain.member.entity.QMember.member;
 import static site.timecapsulearchive.core.domain.member_group.entity.QMemberGroup.memberGroup;
@@ -92,25 +90,10 @@ public class GroupQueryRepositoryImpl implements GroupQueryRepository {
         );
     }
 
-    public Long findGroupCapsuleCount(final Long groupId) {
-        return jpaQueryFactory.select(capsule.count())
-            .from(capsule)
-            .where(capsule.group.id.eq(groupId))
-            .fetchOne();
-    }
-
     public Boolean findGroupEditPermission(final Long groupId, final Long memberId) {
         return jpaQueryFactory.select(memberGroup.isOwner)
             .from(memberGroup)
             .where(memberGroup.group.id.eq(groupId).and(memberGroup.member.id.eq(memberId)))
             .fetchOne();
-    }
-
-    public List<Long> findFriendIds(final List<Long> groupMemberIds, final Long memberId) {
-        return jpaQueryFactory.select(memberFriend.friend.id)
-            .from(memberFriend)
-            .where(
-                memberFriend.friend.id.in(groupMemberIds).and(memberFriend.owner.id.eq(memberId)))
-            .fetch();
     }
 }
