+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group/service/query/GroupQueryService.java
@@ -6,6 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import site.timecapsulearchive.core.domain.capsule.group_capsule.repository.GroupCapsuleQueryRepository;
+import site.timecapsulearchive.core.domain.friend.repository.member_friend.MemberFriendRepository;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupDetailDto;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupDetailTotalDto;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupMemberDto;
@@ -20,6 +22,8 @@ import site.timecapsulearchive.core.domain.group.repository.GroupRepository;
 public class GroupQueryService {
 
     private final GroupRepository groupRepository;
+    private final GroupCapsuleQueryRepository groupCapsuleQueryRepository;
+    private final MemberFriendRepository memberFriendRepository;
 
     public Group findGroupById(final Long groupId) {
         return groupRepository.findGroupById(groupId)
@@ -38,13 +42,13 @@ public class GroupQueryService {
         final GroupDetailDto groupDetailDto = groupRepository.findGroupDetailByGroupIdAndMemberId(groupId,
             memberId).orElseThrow(GroupNotFoundException::new);
 
-        final Long groupCapsuleCount = groupRepository.findGroupCapsuleCount(groupId);
+        final Long groupCapsuleCount = groupCapsuleQueryRepository.findGroupCapsuleCount(groupId);
         final Boolean canGroupEdit = groupRepository.findGroupEditPermission(groupId, memberId);
 
         final List<Long> groupMemberIds = groupDetailDto.members().stream()
             .map(GroupMemberDto::memberId)
             .toList();
-        final List<Long> friendIds = groupRepository.findFriendIds(groupMemberIds, memberId);
+        final List<Long> friendIds = memberFriendRepository.findFriendIds(groupMemberIds, memberId);
 
         return GroupDetailTotalDto.as(groupDetailDto, groupCapsuleCount, canGroupEdit, friendIds);
     }
